intellieye/requirements.txt
transformers>=4.40.0
torch>=2.2.0
torchvision>=0.17.0
accelerate>=0.27.0
pyautogui>=0.9.54
mss>=9.0.1
Pillow>=10.2.0
opencv-python>=4.9.0
pynput>=1.7.6
numpy>=1.26.0
bitsandbytes>=0.43.0
